requirements.txt
python-dotenv==1.0.0
streamlit==1.34.0
pypdf==4.2.0
openai==1.26.0
langchain-openai==0.1.6
huggingface-hub==0.23.0
langchain==0.1.17